Update pv_excess_control.py to improve time trigger management
Old code problems to roubstly have running a time_triggered instance of on_time() running. Often there was no instance and the user was also not able to understand subsequent issues.
To fix this the multiplicty of instance copies in a global array, a local trigger variable was removed. The logics of trying to have just one on_time() method running was changed. The control code now gets only executed for the first instance in the dictionary, but all automations have a on_time() instance. This seems robust.